requirements.txt
dotenv
openai
google-genai
prompt_toolkit
pydantic
jinja2
lxml
rich
